# fix(db): add index for isSyncJobRunning