--- a/packages/shared/lib/db/migrations/20240304130200_jobs-index-order.cjs
+++ b/packages/shared/lib/db/migrations/20240304130200_jobs-index-order.cjs
@@ -1,0 +1,11 @@
+exports.config = { transaction: false };
+
+exports.up = async function (knex) {
+    await knex.schema.raw(
+        'CREATE INDEX CONCURRENTLY "idx_jobs_syncid_createdat_where_deleted" ON "_nango_sync_jobs" USING BTREE ("sync_id", "created_at" DESC) WHERE deleted = false'
+    );
+};
+
+exports.down = async function (knex) {
+    await knex.schema.raw('DROP INDEX CONCURRENTLY idx_jobs_syncid_createdat_where_deleted');
+};